Quick fix update
NOTE: This is a quick update to make the mod usable as before, but I
would like to completely rewrite it and make it better, not sure when
I'll finish that however.

- Workaround for no interaction through helmet by re-creating the helmet
as a sprite CPU particle effect.
- Added "VisorScale" in settings, making the old "Scale" value obsolete
but still kept there for now, until I rewrite the mod, then the
VisorScale behavior will be used (FOV independent scaling).
- Changed "/helmet scale" command to read/write the VisorScale setting.
- Fixed display not showing equipped rifle stats.

Changes from a previous version:
- Fixed compile issues due to radio antenna changes.
- Fixed LCD emissivity due to LCD changes.